+++ src/corpora/Library.vue
@@ -1,11 +1,12 @@
 <template>
+  <PageTitle>{{ $t("library") }}</PageTitle>
   <Section v-if="isAuthenticated" :title="$t('corpuses')">
     <HelpBox>
       <p>
         {{
           corpusStore.hasCorpora
-            ? $t("dashboard.help.corpora")
-            : $t("dashboard.help.corpora.none")
+            ? $t("library.help.corpora")
+            : $t("library.help.corpora.none")
         }}
       </p>
     </HelpBox>
@@ -40,6 +41,7 @@ import PendingContent from "@/spin/PendingContent.vue";
 import CorpusButton from "./CorpusButton.vue";
 import { useAuth } from "@/auth/auth.composable";
 import SourceUpload from "@/corpus/sources/SourceUpload.vue";
+import PageTitle from "@/components/PageTitle.vue";
 import HelpBox from "@/components/HelpBox.vue";
 import useCorpora from "@/corpora/corpora.composable";
 import { useCorpusStore } from "@/store/corpus.store";